+++ actions/set-mode-profile.js
@@ -1,7 +1,21 @@
 import {config} from '../config.js';
+import {
+	getCurrentMode,
+	isAutomaticSwitchingEnabled,
+	isTerminalOpen,
+	setTerminalProfile,
+} from '../functions/index.js';
 
-export function setModeProfile({mode, profile}) {
+export async function setModeProfile({mode, profile}) {
 	config[`${mode}Profile`] = profile;
+
+	if ((await isAutomaticSwitchingEnabled()) && (await isTerminalOpen())) {
+		const currentMode = await getCurrentMode();
+
+		if (currentMode === mode) {
+			await setTerminalProfile(profile);
+		}
+	}
 
 	console.log(`${mode} mode profile set to '${profile}'`);
 }