+++ functions/index.js
@@ -1,5 +1,7 @@
 export {enableAutomaticSwitching} from './enable-automatic-switching.js';
 export {disableAutomaticSwitching} from './disable-automatic-switching.js';
+export {getCurrentMode} from './get-current-mode.js';
 export {getLaunchAgentPlistFileContents} from './get-launch-agent-plist-file-contents.js';
 export {isAutomaticSwitchingEnabled} from './is-automatic-switching-enabled.js';
 export {isTerminalOpen} from './is-terminal-open.js';
+export {setTerminalProfile} from './set-terminal-profile.js';