+++ changelog.md
@@ -5,11 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased](https://github.com/patrik-csak/auto-terminal-profile/compare/v4.0.0...HEAD)
+
+### Changed
+
+- **BREAKING**: `enable` and `set-(dark|light)-profile` now set the Terminal profile
+
 ## [4.0.0](https://github.com/patrik-csak/auto-terminal-profile/compare/v3.0.2...v4.0.0) – 2024-03-25
 
 ### Added
 
-- Documentation explaining how to automatically update Terminal profile when opening Terminal 
+- Documentation explaining how to automatically update Terminal profile when opening Terminal
 
 ### Changed